fix verbose print for idle-kicker
some platforms may have a 64bit useconds_t.. since the maximum value we can
possibly sleep is 100m (1m / (1 / 100)), it's easier to cast explicitly to int
for the print. Otherwise I have to change the format and then always upcast it
to 64bit.

+++ memcached.c
@@ -326,7 +326,7 @@ static void *conn_timeout_thread(void *arg) {
             if ((i % CONNS_PER_SLICE) == 0) {
                 if (settings.verbose > 2)
                     fprintf(stderr, "idle timeout thread sleeping for %ulus\n",
-                        timeslice);
+                        (unsigned int)timeslice);
                 usleep(timeslice);
             }
